Move represents type
