+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/declaration/kocomplexdeclaration/snippet/forclass/class-represents-type.kttxt
