+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/declaration/kocomplexdeclaration/snippet/forcompanionobject/companion-object-represents-type.kttxt
