+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/declaration/kocomplexdeclaration/snippet/forinterface/interface-represents-type.kttxt
