+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/declaration/kocomplexdeclaration/snippet/forobject/object-represents-type.kttxt
