+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/declaration/kodeclaration/KoDeclarationExtForCompanionObjectTest.kt
@@ -2,12 +2,9 @@ package com.lemonappdev.konsist.api.ext.declaration.kodeclaration
 
 import com.lemonappdev.konsist.TestSnippetProvider
 import com.lemonappdev.konsist.api.ext.declaration.hasAnnotationOf
-import com.lemonappdev.konsist.api.ext.declaration.representsTypeOf
 import com.lemonappdev.konsist.testdata.NonExistingAnnotation
 import com.lemonappdev.konsist.testdata.SampleAnnotation1
 import com.lemonappdev.konsist.testdata.SampleAnnotation2
-import com.lemonappdev.konsist.testdata.SampleTopLevelInterface.SampleCompanionObject
-import com.lemonappdev.konsist.testdata.SampleType
 import org.amshove.kluent.assertSoftly
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
@@ -25,22 +22,6 @@ class KoDeclarationExtForCompanionObjectTest {
             hasAnnotationOf<SampleAnnotation1>() shouldBeEqualTo true
             hasAnnotationOf<SampleAnnotation2>() shouldBeEqualTo true
             hasAnnotationOf<NonExistingAnnotation>() shouldBeEqualTo false
-        }
-    }
-
-    @Test
-    fun `companion-object-represents-type`() {
-        // given
-        val sut = getSnippetFile("companion-object-represents-type")
-            .interfaces()
-            .first()
-            .companionObjects()
-            .first()
-
-        // then
-        assertSoftly(sut) {
-            representsTypeOf<SampleCompanionObject>() shouldBeEqualTo true
-            representsTypeOf<SampleType>() shouldBeEqualTo false
         }
     }
