+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/api/ext/declaration/kodeclaration/KoDeclarationExtForObjectTest.kt
@@ -2,12 +2,9 @@ package com.lemonappdev.konsist.api.ext.declaration.kodeclaration
 
 import com.lemonappdev.konsist.TestSnippetProvider
 import com.lemonappdev.konsist.api.ext.declaration.hasAnnotationOf
-import com.lemonappdev.konsist.api.ext.declaration.representsTypeOf
 import com.lemonappdev.konsist.testdata.NonExistingAnnotation
 import com.lemonappdev.konsist.testdata.SampleAnnotation1
 import com.lemonappdev.konsist.testdata.SampleAnnotation2
-import com.lemonappdev.konsist.testdata.SampleObject
-import com.lemonappdev.konsist.testdata.SampleType
 import org.amshove.kluent.assertSoftly
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
@@ -25,20 +22,6 @@ class KoDeclarationExtForObjectTest {
             hasAnnotationOf<SampleAnnotation1>() shouldBeEqualTo true
             hasAnnotationOf<SampleAnnotation2>() shouldBeEqualTo true
             hasAnnotationOf<NonExistingAnnotation>() shouldBeEqualTo false
-        }
-    }
-
-    @Test
-    fun `object-represents-type`() {
-        // given
-        val sut = getSnippetFile("object-represents-type")
-            .objects()
-            .first()
-
-        // then
-        assertSoftly(sut) {
-            representsTypeOf<SampleObject>() shouldBeEqualTo true
-            representsTypeOf<SampleType>() shouldBeEqualTo false
         }
     }
